third_party/blink/web_tests/fast/forms/week/week-input-type-expected.txt

Tests for writing and reading .type property of HTMLInputElement.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS input.type for "week" is correctly "week".
PASS input.type for "WEEK" is correctly "week".
PASS input.type for " week " is correctly "text".
PASS successfullyParsed is true

TEST COMPLETE
